chore: Update initAlias.sh script
This commit adds new aliases to the initAlias.sh script. The aliases include shortcuts for poweroff, speedtest-cli, and opening the Brave browser. Additionally, it includes aliases for file and directory navigation, running development commands with pnpm, updating and installing packages with dnf, and more. This update improves the functionality and convenience of the script.